move code around and remove scalar fn plugin
