+++ vortex-turboquant/src/scalar_fns/decode.rs
@@ -7,39 +7,43 @@ use std::fmt;
 use std::fmt::Formatter;
 use std::sync::Arc;
 
+use num_traits::Float;
+use num_traits::FromPrimitive;
 use vortex_array::ArrayRef;
 use vortex_array::ExecutionCtx;
+use vortex_array::IntoArray;
+use vortex_array::arrays::FixedSizeListArray;
+use vortex_array::arrays::PrimitiveArray;
 use vortex_array::arrays::ScalarFnArray;
-use vortex_array::arrays::scalar_fn::ScalarFnArrayView;
-use vortex_array::arrays::scalar_fn::plugin::ScalarFnArrayParts;
-use vortex_array::arrays::scalar_fn::plugin::ScalarFnArrayVTable;
 use vortex_array::dtype::DType;
+use vortex_array::dtype::NativePType;
 use vortex_array::dtype::Nullability;
 use vortex_array::dtype::extension::ExtDType;
 use vortex_array::expr::Expression;
 use vortex_array::extension::EmptyMetadata;
+use vortex_array::match_each_float_ptype;
 use vortex_array::scalar_fn::Arity;
 use vortex_array::scalar_fn::ChildName;
 use vortex_array::scalar_fn::ExecutionArgs;
 use vortex_array::scalar_fn::ScalarFnId;
 use vortex_array::scalar_fn::ScalarFnVTable;
 use vortex_array::scalar_fn::TypedScalarFnInstance;
-use vortex_array::serde::ArrayChildren;
+use vortex_array::validity::Validity;
+use vortex_buffer::BufferMut;
+use vortex_error::VortexExpect;
 use vortex_error::VortexResult;
-use vortex_error::vortex_ensure_eq;
+use vortex_error::vortex_ensure;
 use vortex_error::vortex_err;
+use vortex_mask::Mask;
 use vortex_session::VortexSession;
-use vortex_tensor::vector::AnyVector;
 use vortex_tensor::vector::Vector;
 
-use super::metadata::deserialize_config;
-use super::metadata::serialize_config;
-use crate::TurboQuantConfig;
-use crate::vector::decode::decode_vector;
-use crate::vtable::TurboQuant;
+use crate::centroids::compute_or_get_centroids;
+use crate::sorf::SorfMatrix;
+use crate::vector::storage::parse_storage;
+use crate::vector::tq_padded_dim;
 use crate::vtable::TurboQuantMetadata;
 use crate::vtable::tq_metadata;
-use crate::vtable::tq_storage_dtype;
 
 /// Lazy TurboQuant vector decode scalar function.
 #[derive(Clone)]
@@ -47,29 +51,26 @@ pub struct TQDecode;
 
 impl TQDecode {
     /// Creates a new [`TypedScalarFnInstance`] wrapping TurboQuant decoding.
-    pub fn new(config: &TurboQuantConfig) -> TypedScalarFnInstance<TQDecode> {
-        TypedScalarFnInstance::new(TQDecode, config.clone())
+    pub fn new() -> TypedScalarFnInstance<TQDecode> {
+        TypedScalarFnInstance::new(TQDecode, EmptyMetadata)
     }
 
     /// Constructs a [`ScalarFnArray`] that lazily decodes a `TurboQuant` child into a `Vector`.
-    pub fn try_new_array(
-        child: ArrayRef,
-        config: &TurboQuantConfig,
-        len: usize,
-    ) -> VortexResult<ScalarFnArray> {
-        ScalarFnArray::try_new(TQDecode::new(config).erased(), vec![child], len)
+    pub fn try_new_array(child: ArrayRef) -> VortexResult<ScalarFnArray> {
+        let len = child.len();
+        ScalarFnArray::try_new(TQDecode::new().erased(), vec![child], len)
     }
 }
 
 impl ScalarFnVTable for TQDecode {
-    type Options = TurboQuantConfig;
+    type Options = EmptyMetadata;
 
     fn id(&self) -> ScalarFnId {
         ScalarFnId::new("vortex.turboquant.decode")
     }
 
-    fn serialize(&self, options: &Self::Options) -> VortexResult<Option<Vec<u8>>> {
-        Ok(Some(serialize_config(options)))
+    fn serialize(&self, _options: &Self::Options) -> VortexResult<Option<Vec<u8>>> {
+        Ok(Some(vec![]))
     }
 
     fn deserialize(
@@ -77,7 +78,12 @@ impl ScalarFnVTable for TQDecode {
         metadata: &[u8],
         _session: &VortexSession,
     ) -> VortexResult<Self::Options> {
-        deserialize_config(metadata)
+        vortex_ensure!(
+            metadata.is_empty(),
+            "TQDecode options metadata must be empty"
+        );
+
+        Ok(EmptyMetadata)
     }
 
     fn arity(&self, _options: &Self::Options) -> Arity {
@@ -93,19 +99,18 @@ impl ScalarFnVTable for TQDecode {
 
     fn fmt_sql(
         &self,
-        options: &Self::Options,
+        _options: &Self::Options,
         expr: &Expression,
         f: &mut Formatter<'_>,
     ) -> fmt::Result {
         write!(f, "tq_decode(")?;
         expr.child(0).fmt_sql(f)?;
-        write!(f, ", {options})")
+        write!(f, ")")
     }
 
-    fn return_dtype(&self, options: &Self::Options, arg_dtypes: &[DType]) -> VortexResult<DType> {
+    fn return_dtype(&self, _options: &Self::Options, arg_dtypes: &[DType]) -> VortexResult<DType> {
         let child_dtype = &arg_dtypes[0];
         let metadata = tq_metadata(child_dtype)?;
-        validate_config_matches_metadata(options, &metadata)?;
 
         let storage_dtype = DType::FixedSizeList(
             Arc::new(DType::Primitive(
@@ -146,68 +151,157 @@ impl ScalarFnVTable for TQDecode {
     }
 }
 
-impl ScalarFnArrayVTable for TQDecode {
-    fn serialize(
-        &self,
-        view: &ScalarFnArrayView<Self>,
-        _session: &VortexSession,
-    ) -> VortexResult<Option<Vec<u8>>> {
-        Ok(Some(serialize_config(view.options)))
+/// Decode a `TurboQuant` extension array back into a `Vector` extension array.
+///
+/// The decoded directions are inverse-transformed, truncated to the original dimension, and
+/// multiplied by the stored row norms. The conversion is lossy and does not roundtrip with
+/// [`TQEncode`](crate::TQEncode).
+pub(crate) fn decode_vector(input: ArrayRef, ctx: &mut ExecutionCtx) -> VortexResult<ArrayRef> {
+    let parsed = parse_storage(input, ctx)?;
+    let metadata = parsed.metadata;
+    if parsed.len == 0 {
+        return build_empty_vector(metadata, parsed.vector_validity);
     }
 
-    fn deserialize(
-        &self,
-        dtype: &DType,
-        len: usize,
-        metadata: &[u8],
-        children: &dyn ArrayChildren,
-        _session: &VortexSession,
-    ) -> VortexResult<ScalarFnArrayParts<Self>> {
-        let options = deserialize_config(metadata)?;
-        let vector_metadata = dtype
-            .as_extension_opt()
-            .and_then(|ext_dtype| ext_dtype.metadata_opt::<AnyVector>())
-            .ok_or_else(|| {
-                vortex_err!("TQDecode parent dtype must be a Vector extension array, got {dtype}")
-            })?;
+    let padded_dim = tq_padded_dim(metadata.dimensions)?;
+    let transform = SorfMatrix::try_new(padded_dim, metadata.num_rounds as usize, metadata.seed)?;
+    let padded_dim = u32::try_from(padded_dim)
+        .map_err(|_| vortex_err!("TurboQuant padded dimension does not fit u32"))?;
 
-        let metadata = TurboQuantMetadata {
-            element_ptype: vector_metadata.element_ptype(),
-            dimensions: vector_metadata.dimensions(),
-            bit_width: options.bit_width(),
-            seed: options.seed(),
-            num_rounds: options.num_rounds(),
-        };
-        let storage_dtype = tq_storage_dtype(&metadata, dtype.nullability())?;
-        let child_dtype =
-            DType::Extension(ExtDType::<TurboQuant>::try_new(metadata, storage_dtype)?.erased());
-        let child = children.get(0, &child_dtype, len)?;
+    let centroids = compute_or_get_centroids(padded_dim, metadata.bit_width)?;
 
-        Ok(ScalarFnArrayParts {
-            options,
-            children: vec![child],
-        })
-    }
+    match_each_float_ptype!(metadata.element_ptype, |T| {
+        decode_typed::<T>(
+            DecodeInputs {
+                metadata: &metadata,
+                sorf_matrix: &transform,
+                centroids: &centroids,
+                norms: &parsed.norms,
+                codes: &parsed.codes,
+            },
+            parsed.vector_validity,
+            parsed.len,
+            ctx,
+        )
+    })
 }
 
-fn validate_config_matches_metadata(
-    config: &TurboQuantConfig,
-    metadata: &TurboQuantMetadata,
-) -> VortexResult<()> {
-    vortex_ensure_eq!(
-        config.bit_width(),
-        metadata.bit_width,
-        "TQDecode config bit_width must match TurboQuant child metadata"
-    );
-    vortex_ensure_eq!(
-        config.seed(),
-        metadata.seed,
-        "TQDecode config seed must match TurboQuant child metadata"
-    );
-    vortex_ensure_eq!(
-        config.num_rounds(),
-        metadata.num_rounds,
-        "TQDecode config num_rounds must match TurboQuant child metadata"
-    );
-    Ok(())
+fn build_empty_vector(
+    metadata: TurboQuantMetadata,
+    vector_validity: Validity,
+) -> VortexResult<ArrayRef> {
+    match_each_float_ptype!(metadata.element_ptype, |T| {
+        let elements = PrimitiveArray::empty::<T>(Nullability::NonNullable);
+        let fsl = FixedSizeListArray::try_new(
+            elements.into_array(),
+            metadata.dimensions,
+            vector_validity,
+            0,
+        )?;
+
+        Vector::try_new_vector_array(fsl.into_array())
+    })
+}
+
+struct DecodeInputs<'a> {
+    metadata: &'a TurboQuantMetadata,
+    sorf_matrix: &'a SorfMatrix,
+    centroids: &'a [f32],
+    norms: &'a PrimitiveArray,
+    codes: &'a PrimitiveArray,
+}
+
+fn decode_typed<T>(
+    decode: DecodeInputs<'_>,
+    vector_validity: Validity,
+    num_vectors: usize,
+    ctx: &mut ExecutionCtx,
+) -> VortexResult<ArrayRef>
+where
+    T: NativePType + Float + FromPrimitive,
+{
+    let metadata = decode.metadata;
+    let dimensions = usize::try_from(metadata.dimensions)
+        .vortex_expect("dimensions stays representable as usize");
+    let padded_dim = decode.sorf_matrix.padded_dim();
+    let centroids = decode.centroids;
+    let norms = decode.norms.as_slice::<T>();
+    let codes = decode.codes.as_slice::<u8>();
+    let mask = vector_validity.execute_mask(num_vectors, ctx)?;
+
+    let output_len = num_vectors
+        .checked_mul(dimensions)
+        .ok_or_else(|| vortex_err!("TurboQuant decoded vector length overflow"))?;
+    let mut output = BufferMut::<T>::with_capacity(output_len);
+
+    let mut decoded = vec![0.0f32; padded_dim];
+    let mut inverse = vec![0.0f32; padded_dim];
+
+    let mut decode_row = |output: &mut BufferMut<T>, i: usize| {
+        let code_row = &codes[i * padded_dim..][..padded_dim];
+
+        for (dst, &code) in decoded.iter_mut().zip(code_row.iter()) {
+            *dst = *centroids
+                .get(usize::from(code))
+                .vortex_expect("TurboQuant code exceeds centroid count");
+        }
+
+        decode.sorf_matrix.inverse_transform(&decoded, &mut inverse);
+
+        let norm = norms[i];
+        for &value in inverse.iter().take(dimensions) {
+            // `T::from_f32` is infallible for the supported float ptypes (`f16`, `f32`,
+            // `f64`): values outside `f16` range saturate to `±inf` rather than returning
+            // `None`.
+            let value = T::from_f32(value)
+                .vortex_expect("from_f32 is infallible for supported float types");
+
+            // SAFETY: total pushes across all match arms equal `output_len`.
+            unsafe { output.push_unchecked(value * norm) };
+        }
+    };
+
+    match &mask {
+        Mask::AllFalse(_) => {
+            // SAFETY: `output` was allocated with capacity `output_len`, and this push writes
+            // exactly `output_len` zero placeholders.
+            unsafe { output.push_n_unchecked(T::zero(), output_len) };
+        }
+        Mask::AllTrue(_) => {
+            for i in 0..num_vectors {
+                decode_row(&mut output, i);
+            }
+        }
+        Mask::Values(values_mask) => {
+            let mut cursor = 0;
+
+            for &(start, end) in values_mask.slices() {
+                if start > cursor {
+                    // SAFETY: total pushes across all arms equal `output_len`.
+                    unsafe { output.push_n_unchecked(T::zero(), (start - cursor) * dimensions) };
+                }
+
+                for i in start..end {
+                    decode_row(&mut output, i);
+                }
+
+                cursor = end;
+            }
+
+            if cursor < num_vectors {
+                // SAFETY: total pushes across all arms equal `output_len`.
+                unsafe { output.push_n_unchecked(T::zero(), (num_vectors - cursor) * dimensions) };
+            }
+        }
+    }
+
+    let elements = PrimitiveArray::new::<T>(output.freeze(), Validity::NonNullable);
+    let fsl = FixedSizeListArray::try_new(
+        elements.into_array(),
+        metadata.dimensions,
+        vector_validity,
+        num_vectors,
+    )?;
+
+    Vector::try_new_vector_array(fsl.into_array())
 }
